Project/requirements.txt
# Core
numpy>=1.23
pandas>=1.5
scikit-learn>=1.1
matplotlib>=3.6
seaborn>=0.12
opencv-python>=4.7
mediapipe>=0.10.14
# Deep Learning
torch>=2.2
torchvision>=0.17
# Utils
pyyaml>=6.0
tqdm>=4.66
rich>=13.7
